refactor file structure for the application
