+++ src/components/messages/flasshMessagesList.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {connect} from 'react-redux';
 import  PropTypes from 'prop-types';
 import FlashMessage from './FlashMessage';
-import {deleteFlashMessage} from '../../actions/flashMessages';
+import {deleteFlashMessage} from '../../redux/actions/flashMessages';
 
 class FlashMessagesList extends React.Component {
     render(){
